requirements.txt
numpy>=1.23
Pillow>=10.0
matplotlib>=3.8
torch>=2.2
torchvision>=0.17
tqdm>=4.66
scikit-image>=0.23
